Fix issue with form Vue components when there are several forms on the page
There was an issue when there were multiple forms on a page with
data['fields'], since the 'field-mount' event was emitted from the
root, the forms were recording every field on the page in data['fields']
which resulted in issues in saving the form. This issue was fixed by
checking on form Vue components that the parent_id passed in the
field-mount event is the forms uid or the uid of one of the forms child
components.