Remove invalid argument to Formatter constructor
In Python 3.7, the first argument to logging.Formatter.__init__ is not
validated by default; however, in Python 3.8 and newer, the fmt string
is validated. Since `self` is incorrectly being passed as the value for
the `fmt` parameter, this results in a TypeError when the validation
happens since `JsonFormatter` is not a `str`. Passing `self` seems to be
a mistake.