Fix ops Docker container python symlink
The previous symlink created a link from a non-existent target to a
misleading name. It seems that perhaps the intent was to make it so
invoking `python` invoked the `python3.7` binary; however:
  - Python 3.7.3's `make install` places data in /usr/local
  - The arguments are backwards so python3.7 is a broken link
  - Neither the base RHEL install or `make install` create a file in
    /usr/bin nor /usr/local/bin with the name `python`

Flipping the arguments, using `/usr/local/` and now referencing
`python3.8` makes a `python` executable available on the PATH; however,
it is unclear whether this is actually needed by any scripts since it
would not have worked before.