Bump pip version to 21.*
In 2018, pip switched to using date-based versioning rather than
semantic versioning, so the version jumped from 10 to 18. `pip` 20.3
switched to the new dependency resolver; this does not seem to cause any
issues with our `setup` script or dependency resolution.

This jump is therefore not quite as signficant as it may seem and
keeping pip up-to-date with the Python release is a worthwhile
endeavor.