Improve NSG flow log configuration
Many of the uses of this want the resource ID instead of the workspace
ID. Additionally, there are some invalid port numbers. Using the default
NetworkWatcher, which is created automatically in every subscription in
every region (and for which there is a limit of 1 for the same),
prevents error messages about duplicate watchers.