Remove ops/requirements.txt
This folder is no longer necessary with the upcoming
`atat-infrastructure` work but we don't want to remove the entire folder
until the ticket can be implemented thoughtfully and in earnest. Since
the `requirements.txt` already has conflicting dependencies and a lack
of documentation, deleting it to avoid dependabot PRs that will cause
further breakage and noise may be the best option.

This is a compromise solution until we can remove the `ops/` folder
completely.