Update "Sonar Scans" GHA name (#2249)
I don't want to bikeshed over the naming here too much but I think that
we're running into confusion over what exactly the goal of the pipeline
does. We do quite a bit more than just a static analysis of the code.
Over time, of course we want to make this better. But today, we can at
least be clearer about exactly what's happening. Perhaps that will
reduce some of the complexity to understanding failures and where tests
should go.