Use `date-fns` for nearly all date calculations
Using a switch based on the month is unnecessary when February is the
only month that changes. This code is largely intended, it seems, to
properly handle date calculations when `this.year` is `null` (or
falsey); however, `new Date(null, this.month)` would yield a month in
the year 1900. This seems fine for calulating non-February. For February
it does cause an issue but continuing to special case that and returning
29 keeps with the current spirit of the code (and keeps the existing
test cases passing).