Make a basic attempt at preventing file overwrite
As mentioned in the comment, this is potentially susceptible to TOCTOU
in rare cases but it _is_ an improvement and prevents the major concern
of overwriting files that already exist. It does not necessarily prevent
the scenario where we issue two clients the same blob name before either
uploads. Paired with UUID4 this should be exceptionally rare.