Prevent using Mock upload in ea-hybrid
This code was not adapted for hybrid environments that are not names
specifically "hybrid". As a slight improvement, if an environment
contains "hybrid" at all it is now considered an environment that should
use the Azure storage provider. In the longer term, this logic should be
decoupled from the name of the CSP class in use (for which a ticket
exists).