Improve Task Order PDF uniqueness guarantees (#2271)
* Make a basic attempt at preventing file overwrite

As mentioned in the comment, this is potentially susceptible to TOCTOU
in rare cases but it _is_ an improvement and prevents the major concern
of overwriting files that already exist. It does not necessarily prevent
the scenario where we issue two clients the same blob name before either
uploads. Paired with UUID4 this should be exceptionally rare.

* Remove an incorrect TODO

We absolutely should not perform object name verification client-side
which is untrusted. This would open us up to even more significant
overwrite attempts. Today, the issues are far more theoretical than
practical whereas implementing what the TODO describes would open us up
to far more practical (rather than purely theoretical) concerns.

* Add a unique constraint for PDF attachment IDs

This ensures that the same attachment ID cannot be used for multiple
task orders. If two task orders do have the same PDF, it must be
uploaded separately, with a different ID and object name for each. This
prevents potential issues with overwriting the objects/TO PDFs.

* Prevent using Mock upload in ea-hybrid

This code was not adapted for hybrid environments that are not names
specifically "hybrid". As a slight improvement, if an environment
contains "hybrid" at all it is now considered an environment that should
use the Azure storage provider. In the longer term, this logic should be
decoupled from the name of the CSP class in use (for which a ticket
exists).