Update CI and test workflows
This separation moves a good bit of things that have traditionally been
rather serial in the CI pipeline to being done in parallel. Primarily,
this is the lint, bandit, and formatting checks which have no
dependencies on one another. The SonarCloud scan pipeline happens while
those are happening; eventually, the UI tests will be included in that
too.

This does add a little bit of code duplication across the files in the
.github/workflows/* files. This is unfortunate but there isn't a great
way to share code or steps between the various workflows (or even
between jobs in the same workflow). The improved cleanliness and
feedback largely seems to make up for that deficiency.

This also again changes the name of one of our required branch
protection checks which may prevent this PR from being merged especially
easily; the rules will need to be updated again after the merge as well.