Run UI tests in the CI pipeline
This adds the UI tests to the CI pipeline; modifying various parts of
the codebase to ensure that the tests are run properly and that errors
are displayed correctly in both GitHub Actions and in the BrowserStack
Console.

The most significant change from a user experience perspective is that
the necessary command line arguments for pytest have changed. This is
due to the way that conftest has been modified which allows for
considering the browser and execution backend as two totally separate
actions. The documentation will need to be updated to reflect this.

This also adds support for the BrowserStackLocal execution method, which
removes the need to run ngrok while still letting
non-Internet-accessible sites (such as the one on localhost:8000) to be
tested easily. BrowserStack very cleanly supports this, and the
necessary environment variables have been added to the GitHub repository
already.

Futher, a few modifications have been made to nearly all of the UI test
functions; and a nearly-transparent change has been made to support
future tests. First, a BrowserStack utility module has been added.
This provides two features:
 1. Easy functions for commonly-used API actions
 2. A simple wrapper class to support error handling
All test functions have been modified to use the corresponding API
functions. Additionally, the new wrapper class is returned from the
`setup` fixture rather than a traditional Selenium WebDriver when using
BrowserStack. This should be entirely transparent to those tests. The
new wrapper will send a setSessionStatus API call to mark a test as
failed when there is any exception in a Selenium WebDriver function
call. The same error will still get raised again; however, to allow the
higher-level tests to change the status again (either at that site or
another) to ensure statuses are properly reported.

Finally, since the `setup` fixture `yield`s the Driver from within a
context manager (`with`), the `quit()` method no longer needs to be
called by each test (nor should it).

Hopefully altogether, these not only make it possible to run the UI
tests in the CI pipeline but also make it easier to write more tests
in the future.