Fix SonarCloud PR reporting
When run outside the `pull_request` context, the variables that Sonar
Scanner uses to determine the pull request information are not available
in the `github` context object. Those need to be provided manually and
must be passed from the earlier workflow stage. They can be provided in
the output artifact in the same way that the ref and coverage reports
are.